End Week 02

+++ OrdersConsoleApp/MenuAction.cs
@@ -1,15 +1,14 @@
 ﻿namespace OrdersConsoleApp;
-
 public class MenuAction
 {
-    public MenuAction(int id, string name, string menuName)
+    public MenuAction(byte id, string name, string menuName)
     {
         Id = id;
         Name = name;
         MenuName = menuName;
     }
 
-    public int Id { get; set; }
+    public byte Id { get; set; }
     public string Name { get; set; }
     public string MenuName { get; set; }
 }
